adapter: move hydration check to dedicated module
We'll be adding more logic to this module shortly, so this commit
is preparation to keep all that logic nicely boxed up.

Pure code movement. No behavior changes.

Touches #27413.
